add empty audio channel in relay

+++ src/node_trans_session.js
@@ -81,13 +81,18 @@ class NodeTransSession extends EventEmitter {
       fs.readdir(ouPath, function (err, files) {
         if (!err) {
           files.forEach((filename) => {
-            if ((filename.endsWith('.ts')
-              || filename.endsWith('.m3u8')
-              || filename.endsWith('.mpd')
-              || filename.endsWith('.m4s')
-              || filename.endsWith('.tmp'))
-              && fs.existsSync(ouPath + '/' + filename)) {
-              fs.unlinkSync(ouPath + '/' + filename);
+            if ((
+                filename.endsWith('.ts')
+                || filename.endsWith('.m3u8')
+                || filename.endsWith('.mpd')
+                || filename.endsWith('.m4s')
+                || filename.endsWith('.tmp')
+              ) && fs.existsSync(ouPath + '/' + filename)
+              && !this.conf.noDelete
+            ) {
+              try {
+                fs.unlinkSync(ouPath + '/' + filename);
+              } catch(err) {}
             }
           });
         }
